Move AudienceSection to shared paid ads component.
