+++ js/src/setup-ads/ads-stepper/create-campaign/form-content.js
@@ -1,8 +1,13 @@
 /**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
  * Internal dependencies
  */
 import StepContentFooter from '.~/components/stepper/step-content-footer';
-import AudienceSection from './audience-section';
+import AudienceSection from '.~/components/paid-ads/audience-section';
 import BudgetSection from './budget-section';
 import FaqsSection from './faqs-section';
 
@@ -11,7 +16,13 @@ const FormContent = ( props ) => {
 
 	return (
 		<>
-			<AudienceSection formProps={ formProps } />
+			<AudienceSection
+				formProps={ formProps }
+				countrySelectHelperText={ __(
+					'You can only select one country per campaign. ',
+					'google-listings-and-ads'
+				) }
+			/>
 			<BudgetSection formProps={ formProps } />
 			<FaqsSection />
 			<StepContentFooter>{ submitButton }</StepContentFooter>
